+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Implemented provider of folding ranges for QRC, QBN, message, task and comment blocks.
 - Numeric identifiers of additional messages can now be "renamed". Static messages can't be renamed because id is associated to alias.
 - Added word highlight for symbol types.
+- Added diagnostic for quest names inside quest tables. Quests are seeked in the directory where the table is located.
 
 ### Improved